@@ -16,11 +16,10 @@
         }
     ],
     "require": {
-        "php": "^7.1",
+        "php": ">=7.2",
         "ext-json": "*",
         "beberlei/assert": "^3.2",
         "league/uri": "^6.0",
-        "league/uri-components": "^2.1",
         "psr/http-client": "^1.0",
         "psr/http-factory": "^1.0",
         "psr/log": "^1.1"
@@ -36,13 +35,5 @@
     "suggest": {
         "web-token/jwt-key-mgmt": "Mandatory for fetching Metadata Statement from distant sources",
         "web-token/jwt-signature-algorithm-ecdsa": "Mandatory for fetching Metadata Statement from distant sources"
-    },
-    "extra": {
-        "branch-alias": {
-            "v2.1": "2.1.x-dev",
-            "v3.0": "3.0.x-dev",
-            "v3.1": "3.1.x-dev",
-            "v3.2": "3.2.x-dev"
-        }
     }
 }
