@@ -17,14 +17,14 @@ use Assert\Assertion;
 use Cose\Algorithm\Signature\Signature;
 use Cose\Key\Ec2Key;
 use Cose\Key\Key;
+use function Safe\openssl_sign;
 
 abstract class ECDSA implements Signature
 {
     public function sign(string $data, Key $key): string
     {
         $key = $this->handleKey($key);
-        $result = openssl_sign($data, $signature, $key->asPEM(), $this->getHashAlgorithm());
-        Assertion::true($result, 'Unable to sign the data');
+        openssl_sign($data, $signature, $key->asPEM(), $this->getHashAlgorithm());
 
         return ECSignature::fromAsn1($signature, $this->getSignaturePartLength());
     }
@@ -38,6 +38,12 @@ abstract class ECDSA implements Signature
         return 1 === openssl_verify($data, $signature, $publicKey->asPEM(), $this->getHashAlgorithm());
     }
 
+    abstract protected function getCurve(): int;
+
+    abstract protected function getHashAlgorithm(): int;
+
+    abstract protected function getSignaturePartLength(): int;
+
     private function handleKey(Key $key): Ec2Key
     {
         $key = new Ec2Key($key->getData());
@@ -45,10 +51,4 @@ abstract class ECDSA implements Signature
 
         return $key;
     }
-
-    abstract protected function getCurve(): int;
-
-    abstract protected function getHashAlgorithm(): int;
-
-    abstract protected function getSignaturePartLength(): int;
 }
