@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Cose\Algorithm\Signature\RSA;
 
 use function ceil;
+use function chr;
 use Cose\Algorithm\Signature\Signature;
 use Cose\Key\Key;
 use Cose\Key\RsaKey;
@@ -23,9 +24,10 @@ use Jose\Component\Core\Util\BigInteger;
 use Jose\Component\Core\Util\Hash;
 use function mb_strlen;
 use function mb_substr;
-use function pack;
+use function ord;
 use function random_bytes;
 use RuntimeException;
+use function Safe\pack;
 use function str_pad;
 use function str_repeat;
 
@@ -62,12 +64,38 @@ abstract class PSSRSA implements Signature
         return $this->verifyEMSAPSS($data, $em, $modBits - 1, $this->getHashAlgorithm());
     }
 
+    /**
+     * Exponentiate with or without Chinese Remainder Theorem.
+     * Operation with primes 'p' and 'q' is appox. 2x faster.
+     */
+    public function exponentiate(RsaKey $key, BigInteger $c): BigInteger
+    {
+        if ($c->compare(BigInteger::createFromDecimal(0)) < 0 || $c->compare(BigInteger::createFromBinaryString($key->n())) > 0) {
+            throw new RuntimeException();
+        }
+        if ($key->isPublic() || !$key->hasPrimes() || !$key->hasExponents() || !$key->hasCoefficient()) {
+            return $c->modPow(BigInteger::createFromBinaryString($key->e()), BigInteger::createFromBinaryString($key->n()));
+        }
+
+        $p = $key->primes()[0];
+        $q = $key->primes()[1];
+        $dP = $key->exponents()[0];
+        $dQ = $key->exponents()[1];
+        $qInv = BigInteger::createFromBinaryString($key->QInv());
+
+        $m1 = $c->modPow($dP, $p);
+        $m2 = $c->modPow($dQ, $q);
+        $h = $qInv->multiply($m1->subtract($m2)->add($p))->mod($p);
+
+        return $m2->add($h->multiply($q));
+    }
+
+    abstract protected function getHashAlgorithm(): Hash;
+
     private function handleKey(Key $key): RsaKey
     {
         return new RsaKey($key->getData());
     }
-
-    abstract protected function getHashAlgorithm(): Hash;
 
     private function convertIntegerToOctetString(BigInteger $x, int $xLen): string
     {
@@ -76,7 +104,7 @@ abstract class PSSRSA implements Signature
             throw new RuntimeException('Unable to convert the integer');
         }
 
-        return str_pad($x, $xLen, \chr(0), STR_PAD_LEFT);
+        return str_pad($x, $xLen, chr(0), STR_PAD_LEFT);
     }
 
     /**
@@ -108,13 +136,13 @@ abstract class PSSRSA implements Signature
         $salt = random_bytes($sLen);
         $m2 = "\0\0\0\0\0\0\0\0".$mHash.$salt;
         $h = $hash->hash($m2);
-        $ps = str_repeat(\chr(0), $emLen - $sLen - $hash->getLength() - 2);
-        $db = $ps.\chr(1).$salt;
+        $ps = str_repeat(chr(0), $emLen - $sLen - $hash->getLength() - 2);
+        $db = $ps.chr(1).$salt;
         $dbMask = $this->getMGF1($h, $emLen - $hash->getLength() - 1, $hash);
         $maskedDB = $db ^ $dbMask;
-        $maskedDB[0] = ~\chr(0xFF << ($modulusLength & 7)) & $maskedDB[0];
+        $maskedDB[0] = ~chr(0xFF << ($modulusLength & 7)) & $maskedDB[0];
 
-        return $maskedDB.$h.\chr(0xBC);
+        return $maskedDB.$h.chr(0xBC);
     }
 
     /**
@@ -128,23 +156,23 @@ abstract class PSSRSA implements Signature
         if ($emLen < $hash->getLength() + $sLen + 2) {
             throw new InvalidArgumentException();
         }
-        if ($em[mb_strlen($em, '8bit') - 1] !== \chr(0xBC)) {
+        if ($em[mb_strlen($em, '8bit') - 1] !== chr(0xBC)) {
             throw new InvalidArgumentException();
         }
         $maskedDB = mb_substr($em, 0, -$hash->getLength() - 1, '8bit');
         $h = mb_substr($em, -$hash->getLength() - 1, $hash->getLength(), '8bit');
-        $temp = \chr(0xFF << ($emBits & 7));
+        $temp = chr(0xFF << ($emBits & 7));
         if ((~$maskedDB[0] & $temp) !== $temp) {
             throw new InvalidArgumentException();
         }
         $dbMask = $this->getMGF1($h, $emLen - $hash->getLength() - 1, $hash/*MGF*/);
         $db = $maskedDB ^ $dbMask;
-        $db[0] = ~\chr(0xFF << ($emBits & 7)) & $db[0];
+        $db[0] = ~chr(0xFF << ($emBits & 7)) & $db[0];
         $temp = $emLen - $hash->getLength() - $sLen - 2;
-        if (mb_substr($db, 0, $temp, '8bit') !== str_repeat(\chr(0), $temp)) {
+        if (mb_substr($db, 0, $temp, '8bit') !== str_repeat(chr(0), $temp)) {
             throw new InvalidArgumentException();
         }
-        if (1 !== \ord($db[$temp])) {
+        if (1 !== ord($db[$temp])) {
             throw new InvalidArgumentException();
         }
         $salt = mb_substr($db, $temp + 1, null, '8bit'); // should be $sLen long
@@ -152,31 +180,5 @@ abstract class PSSRSA implements Signature
         $h2 = $hash->hash($m2);
 
         return hash_equals($h, $h2);
-    }
-
-    /**
-     * Exponentiate with or without Chinese Remainder Theorem.
-     * Operation with primes 'p' and 'q' is appox. 2x faster.
-     */
-    public function exponentiate(RsaKey $key, BigInteger $c): BigInteger
-    {
-        if ($c->compare(BigInteger::createFromDecimal(0)) < 0 || $c->compare(BigInteger::createFromBinaryString($key->n())) > 0) {
-            throw new RuntimeException();
-        }
-        if ($key->isPublic() || !$key->hasPrimes() || !$key->hasExponents() || !$key->hasCoefficient()) {
-            return $c->modPow(BigInteger::createFromBinaryString($key->e()), BigInteger::createFromBinaryString($key->n()));
-        }
-
-        $p = $key->primes()[0];
-        $q = $key->primes()[1];
-        $dP = $key->exponents()[0];
-        $dQ = $key->exponents()[1];
-        $qInv = BigInteger::createFromBinaryString($key->QInv());
-
-        $m1 = $c->modPow($dP, $p);
-        $m2 = $c->modPow($dQ, $q);
-        $h = $qInv->multiply($m1->subtract($m2)->add($p))->mod($p);
-
-        return $m2->add($h->multiply($q));
     }
 }
