@@ -17,6 +17,7 @@ use Assert\Assertion;
 use Cose\Algorithm\Signature\Signature;
 use Cose\Key\Key;
 use Cose\Key\RsaKey;
+use function Safe\openssl_sign;
 
 abstract class RSA implements Signature
 {
@@ -25,8 +26,7 @@ abstract class RSA implements Signature
         $key = $this->handleKey($key);
         Assertion::true($key->isPrivate(), 'The key is not private');
 
-        $result = openssl_sign($data, $signature, $key->asPem(), $this->getHashAlgorithm());
-        Assertion::true($result, 'Unable to sign the data');
+        openssl_sign($data, $signature, $key->asPem(), $this->getHashAlgorithm());
 
         return $signature;
     }
@@ -38,10 +38,10 @@ abstract class RSA implements Signature
         return 1 === openssl_verify($data, $signature, $key->asPem(), $this->getHashAlgorithm());
     }
 
+    abstract protected function getHashAlgorithm(): int;
+
     private function handleKey(Key $key): RsaKey
     {
         return new RsaKey($key->getData());
     }
-
-    abstract protected function getHashAlgorithm(): int;
 }
