@@ -13,8 +13,12 @@ declare(strict_types=1);
 
 namespace Webauthn\MetadataService;
 
+use Assert\Assertion;
+use Base64Url\Base64Url;
 use Psr\Http\Client\ClientInterface;
 use Psr\Http\Message\RequestFactoryInterface;
+use function Safe\json_decode;
+use function Safe\sprintf;
 
 class DistantSingleMetadata extends SingleMetadata
 {
@@ -41,13 +45,13 @@ class DistantSingleMetadata extends SingleMetadata
     /**
      * @var bool
      */
-    private $isBare64Encoded;
+    private $isBase64Encoded;
 
-    public function __construct(string $uri, bool $isBare64Encoded, ClientInterface $httpClient, RequestFactoryInterface $requestFactory, array $additionalHeaders = [])
+    public function __construct(string $uri, bool $isBase64Encoded, ClientInterface $httpClient, RequestFactoryInterface $requestFactory, array $additionalHeaders = [])
     {
-        parent::__construct($uri, $isBare64Encoded); //Useless
+        parent::__construct($uri, $isBase64Encoded); //Useless
         $this->uri = $uri;
-        $this->isBare64Encoded = $isBare64Encoded;
+        $this->isBase64Encoded = $isBase64Encoded;
         $this->httpClient = $httpClient;
         $this->requestFactory = $requestFactory;
         $this->additionalHeaders = $additionalHeaders;
@@ -55,6 +59,24 @@ class DistantSingleMetadata extends SingleMetadata
 
     public function getMetadataStatement(): MetadataStatement
     {
-        return MetadataStatementFetcher::fetchMetadataStatement($this->uri, $this->isBare64Encoded, $this->httpClient, $this->requestFactory, $this->additionalHeaders);
+        $payload = $this->fetch();
+        $json = $this->isBase64Encoded ? Base64Url::decode($payload) : $payload;
+        $data = json_decode($json, true);
+
+        return MetadataStatement::createFromArray($data);
+    }
+
+    private function fetch(): string
+    {
+        $request = $this->requestFactory->createRequest('GET', $this->uri);
+        foreach ($this->additionalHeaders as $k => $v) {
+            $request = $request->withHeader($k, $v);
+        }
+        $response = $this->httpClient->sendRequest($request);
+        Assertion::eq(200, $response->getStatusCode(), sprintf('Unable to contact the server. Response code is %d', $response->getStatusCode()));
+        $content = $response->getBody()->getContents();
+        Assertion::notEmpty($content, 'Unable to contact the server. The response has no content');
+
+        return $content;
     }
 }
