@@ -13,7 +13,8 @@ declare(strict_types=1);
 
 namespace Webauthn\MetadataService;
 
-use Assert\Assertion;
+use function Safe\base64_decode;
+use function Safe\json_decode;
 
 class SingleMetadata
 {
@@ -28,24 +29,22 @@ class SingleMetadata
     /**
      * @var bool
      */
-    private $isBare64Encoded;
+    private $isBase64Encoded;
 
-    public function __construct(string $data, bool $isBare64Encoded)
+    public function __construct(string $data, bool $isBase64Encoded)
     {
         $this->data = $data;
-        $this->isBare64Encoded = $isBare64Encoded;
+        $this->isBase64Encoded = $isBase64Encoded;
     }
 
     public function getMetadataStatement(): MetadataStatement
     {
         if (null === $this->statement) {
             $json = $this->data;
-            if ($this->isBare64Encoded) {
+            if ($this->isBase64Encoded) {
                 $json = base64_decode($this->data, true);
-                Assertion::string($json, 'Unable to decode the data');
             }
             $statement = json_decode($json, true);
-            Assertion::eq(JSON_ERROR_NONE, json_last_error(), 'Unable to decode the data');
             $this->statement = MetadataStatement::createFromArray($statement);
         }
 
