@@ -15,6 +15,7 @@ namespace Webauthn\MetadataService;
 
 use Assert\Assertion;
 use JsonSerializable;
+use function Safe\sprintf;
 
 class VerificationMethodDescriptor implements JsonSerializable
 {
@@ -157,9 +158,9 @@ class VerificationMethodDescriptor implements JsonSerializable
     {
         $data = [
             'userVerification' => $this->userVerification,
-            'caDesc' => $this->caDesc ? $this->caDesc->jsonSerialize() : null,
-            'baDesc' => $this->baDesc ? $this->baDesc->jsonSerialize() : null,
-            'paDesc' => $this->paDesc ? $this->paDesc->jsonSerialize() : null,
+            'caDesc' => null === $this->caDesc ? null : $this->caDesc->jsonSerialize(),
+            'baDesc' => null === $this->baDesc ? null : $this->baDesc->jsonSerialize(),
+            'paDesc' => null === $this->paDesc ? null : $this->paDesc->jsonSerialize(),
         ];
 
         return Utils::filterNullValues($data);
