@@ -13,9 +13,11 @@ declare(strict_types=1);
 
 namespace Webauthn\MetadataService;
 
+use function array_key_exists;
 use Assert\Assertion;
 use JsonSerializable;
 use LogicException;
+use function Safe\sprintf;
 
 class Version implements JsonSerializable
 {
@@ -55,7 +57,7 @@ class Version implements JsonSerializable
     {
         $data = Utils::filterNullValues($data);
         foreach (['major', 'minor'] as $key) {
-            if (\array_key_exists($key, $data)) {
+            if (array_key_exists($key, $data)) {
                 Assertion::integer($data[$key], sprintf('Invalid value for key "%s"', $key));
             }
         }
