@@ -13,7 +13,9 @@ declare(strict_types=1);
 
 namespace Webauthn\AttestationStatement;
 
+use function array_key_exists;
 use Assert\Assertion;
+use function Safe\sprintf;
 
 class AttestationStatementSupportManager
 {
@@ -29,7 +31,7 @@ class AttestationStatementSupportManager
 
     public function has(string $name): bool
     {
-        return \array_key_exists($name, $this->attestationStatementSupports);
+        return array_key_exists($name, $this->attestationStatementSupports);
     }
 
     public function get(string $name): AttestationStatementSupport
