@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace Webauthn\AttestationStatement;
 
+use function array_key_exists;
 use Assert\Assertion;
 use CBOR\Decoder;
 use CBOR\MapObject;
@@ -22,7 +23,9 @@ use Cose\Algorithm\Manager;
 use Cose\Algorithm\Signature\Signature;
 use Cose\Algorithms;
 use Cose\Key\Key;
+use function in_array;
 use InvalidArgumentException;
+use function is_array;
 use RuntimeException;
 use Webauthn\AuthenticatorData;
 use Webauthn\CertificateToolbox;
@@ -56,7 +59,7 @@ final class PackedAttestationStatementSupport implements AttestationStatementSup
     }
 
     /**
-     * @param array<string, mixed> $attestation
+     * @param mixed[] $attestation
      */
     public function load(array $attestation): AttestationStatement
     {
@@ -64,9 +67,9 @@ final class PackedAttestationStatementSupport implements AttestationStatementSup
         Assertion::keyExists($attestation['attStmt'], 'alg', 'The attestation statement value "alg" is missing.');
         Assertion::string($attestation['attStmt']['sig'], 'The attestation statement value "sig" is missing.');
         switch (true) {
-            case \array_key_exists('x5c', $attestation['attStmt']):
+            case array_key_exists('x5c', $attestation['attStmt']):
                 return $this->loadBasicType($attestation);
-            case \array_key_exists('ecdaaKeyId', $attestation['attStmt']):
+            case array_key_exists('ecdaaKeyId', $attestation['attStmt']):
                 return $this->loadEcdaaType($attestation['attStmt']);
             default:
                 return $this->loadEmptyType($attestation);
@@ -89,7 +92,7 @@ final class PackedAttestationStatementSupport implements AttestationStatementSup
     }
 
     /**
-     * @param array<string, mixed> $attestation
+     * @param mixed[] $attestation
      */
     private function loadBasicType(array $attestation): AttestationStatement
     {
@@ -110,7 +113,7 @@ final class PackedAttestationStatementSupport implements AttestationStatementSup
     }
 
     /**
-     * @param array<string, mixed> $attestation
+     * @param mixed[] $attestation
      */
     private function loadEmptyType(array $attestation): AttestationStatement
     {
@@ -129,7 +132,7 @@ final class PackedAttestationStatementSupport implements AttestationStatementSup
         Assertion::false(!isset($parsed['name']) || false === mb_strpos($parsed['name'], '/OU=Authenticator Attestation'), 'Invalid certificate name. The Subject Organization Unit must be "Authenticator Attestation"');
 
         //Check extensions
-        Assertion::false(!isset($parsed['extensions']) || !\is_array($parsed['extensions']), 'Certificate extensions are missing');
+        Assertion::false(!isset($parsed['extensions']) || !is_array($parsed['extensions']), 'Certificate extensions are missing');
 
         //Check certificate is not a CA cert
         Assertion::false(!isset($parsed['extensions']['basicConstraints']) || 'CA:FALSE' !== $parsed['extensions']['basicConstraints'], 'The Basic Constraints extension must have the CA component set to false');
@@ -138,7 +141,7 @@ final class PackedAttestationStatementSupport implements AttestationStatementSup
         Assertion::notNull($attestedCredentialData, 'No attested credential available');
 
         // id-fido-gen-ce-aaguid OID check
-        Assertion::false(\in_array('1.3.6.1.4.1.45724.1.1.4', $parsed['extensions'], true) && !hash_equals($attestedCredentialData->getAaguid()->getBytes(), $parsed['extensions']['1.3.6.1.4.1.45724.1.1.4']), 'The value of the "aaguid" does not match with the certificate');
+        Assertion::false(in_array('1.3.6.1.4.1.45724.1.1.4', $parsed['extensions'], true) && !hash_equals($attestedCredentialData->getAaguid()->getBytes(), $parsed['extensions']['1.3.6.1.4.1.45724.1.1.4']), 'The value of the "aaguid" does not match with the certificate');
     }
 
     private function processWithCertificate(string $clientDataJSONHash, AttestationStatement $attestationStatement, AuthenticatorData $authenticatorData, CertificateTrustPath $trustPath): bool
