@@ -24,9 +24,14 @@ use Cose\Key\Ec2Key;
 use Cose\Key\Key;
 use Cose\Key\OkpKey;
 use Cose\Key\RsaKey;
-use DateTimeImmutable;
+use function count;
+use function in_array;
 use InvalidArgumentException;
+use function is_array;
 use RuntimeException;
+use Safe\DateTimeImmutable;
+use function Safe\sprintf;
+use function Safe\unpack;
 use Webauthn\AuthenticatorData;
 use Webauthn\CertificateToolbox;
 use Webauthn\StringStream;
@@ -41,7 +46,7 @@ final class TPMAttestationStatementSupport implements AttestationStatementSuppor
     }
 
     /**
-     * @param array<string, mixed> $attestation
+     * @param mixed[] $attestation
      */
     public function load(array $attestation): AttestationStatement
     {
@@ -120,7 +125,7 @@ final class TPMAttestationStatementSupport implements AttestationStatementSuppor
     }
 
     /**
-     * @return array<string, mixed>
+     * @return mixed[]
      */
     private function checkCertInfo(string $data): array
     {
@@ -162,7 +167,7 @@ final class TPMAttestationStatementSupport implements AttestationStatementSuppor
     }
 
     /**
-     * @return array<string, mixed>
+     * @return mixed[]
      */
     private function checkPubArea(string $data): array
     {
@@ -195,7 +200,7 @@ final class TPMAttestationStatementSupport implements AttestationStatementSuppor
     }
 
     /**
-     * @return array<string, mixed>
+     * @return mixed[]
      */
     private function getParameters(string $type, StringStream $stream): array
     {
@@ -270,7 +275,7 @@ final class TPMAttestationStatementSupport implements AttestationStatementSuppor
         Assertion::false(!isset($parsed['version']) || 2 !== $parsed['version'], 'Invalid certificate version');
 
         //Check subject field is empty
-        Assertion::false(!isset($parsed['subject']) || !\is_array($parsed['subject']) || 0 !== \count($parsed['subject']), 'Invalid certificate name. The Subject should be empty');
+        Assertion::false(!isset($parsed['subject']) || !is_array($parsed['subject']) || 0 !== count($parsed['subject']), 'Invalid certificate name. The Subject should be empty');
 
         // Check period of validity
         Assertion::keyExists($parsed, 'validFrom_time_t', 'Invalid certificate start date.');
@@ -284,7 +289,7 @@ final class TPMAttestationStatementSupport implements AttestationStatementSuppor
         Assertion::true($endDate > new DateTimeImmutable(), 'Invalid certificate end date.');
 
         //Check extensions
-        Assertion::false(!isset($parsed['extensions']) || !\is_array($parsed['extensions']), 'Certificate extensions are missing');
+        Assertion::false(!isset($parsed['extensions']) || !is_array($parsed['extensions']), 'Certificate extensions are missing');
 
         //Check subjectAltName
         Assertion::false(!isset($parsed['extensions']['subjectAltName']), 'The "subjectAltName" is missing');
@@ -294,9 +299,7 @@ final class TPMAttestationStatementSupport implements AttestationStatementSuppor
         Assertion::eq($parsed['extensions']['extendedKeyUsage'], '2.23.133.8.3', 'The "extendedKeyUsage" is invalid');
 
         // id-fido-gen-ce-aaguid OID check
-        Assertion::false(\in_array('1.3.6.1.4.1.45724.1.1.4', $parsed['extensions'], true) && !hash_equals($authenticatorData->getAttestedCredentialData()->getAaguid()->getBytes(), $parsed['extensions']['1.3.6.1.4.1.45724.1.1.4']), 'The value of the "aaguid" does not match with the certificate');
-
-        // TODO: For attestationRoot in metadata.attestationRootCertificates, generate verification chain verifX5C by appending attestationRoot to the x5c. Try verifying verifX5C. If successful go to next step. If fail try next attestationRoot. If no attestationRoots left to try, fail.
+        Assertion::false(in_array('1.3.6.1.4.1.45724.1.1.4', $parsed['extensions'], true) && !hash_equals($authenticatorData->getAttestedCredentialData()->getAaguid()->getBytes(), $parsed['extensions']['1.3.6.1.4.1.45724.1.1.4']), 'The value of the "aaguid" does not match with the certificate');
     }
 
     private function processWithECDAA(): bool
