@@ -17,6 +17,7 @@ use Assert\Assertion;
 use JsonSerializable;
 use Ramsey\Uuid\Uuid;
 use Ramsey\Uuid\UuidInterface;
+use function Safe\base64_decode;
 
 /**
  * @see https://www.w3.org/TR/webauthn/#sec-attested-credential-data
@@ -66,7 +67,7 @@ class AttestedCredentialData implements JsonSerializable
     }
 
     /**
-     * @param array<string, mixed> $json
+     * @param mixed[] $json
      */
     public static function createFromArray(array $json): self
     {
@@ -78,16 +79,13 @@ class AttestedCredentialData implements JsonSerializable
                 break;
             default: // Kept for compatibility with old format
                 $decoded = base64_decode($json['aaguid'], true);
-                Assertion::string($decoded, 'Unable to decode the data');
                 $uuid = Uuid::fromBytes($decoded);
         }
         $credentialId = base64_decode($json['credentialId'], true);
-        Assertion::string($credentialId, 'Unable to decode the data');
 
         $credentialPublicKey = null;
         if (isset($json['credentialPublicKey'])) {
             $credentialPublicKey = base64_decode($json['credentialPublicKey'], true);
-            Assertion::string($credentialPublicKey, 'Invalid Credential Public Key');
         }
 
         return new self(
@@ -98,7 +96,7 @@ class AttestedCredentialData implements JsonSerializable
     }
 
     /**
-     * @return array<string, mixed>
+     * @return mixed[]
      */
     public function jsonSerialize(): array
     {
