@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Webauthn;
 
-use Assert\Assertion;
+use function Safe\base64_decode;
 
 /**
  * @see https://www.w3.org/TR/webauthn/#authenticatorassertionresponse
@@ -59,9 +59,6 @@ class AuthenticatorAssertionResponse extends AuthenticatorResponse
             return $this->userHandle;
         }
 
-        $decoded = base64_decode($this->userHandle, true);
-        Assertion::string($decoded, 'Unable to decode the data');
-
-        return $decoded;
+        return base64_decode($this->userHandle, true);
     }
 }
