@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace Webauthn;
 
+use function ord;
 use Webauthn\AuthenticationExtensions\AuthenticationExtensionsClientOutputs;
 
 /**
@@ -20,6 +21,12 @@ use Webauthn\AuthenticationExtensions\AuthenticationExtensionsClientOutputs;
  */
 class AuthenticatorData
 {
+    private const FLAG_UP = 0b00000001;
+    private const FLAG_RFU1 = 0b00000010;
+    private const FLAG_UV = 0b00000100;
+    private const FLAG_RFU2 = 0b00111000;
+    private const FLAG_AT = 0b01000000;
+    private const FLAG_ED = 0b10000000;
     /**
      * @var string
      */
@@ -50,13 +57,6 @@ class AuthenticatorData
      */
     protected $extensions;
 
-    private const FLAG_UP = 0b00000001;
-    private const FLAG_RFU1 = 0b00000010;
-    private const FLAG_UV = 0b00000100;
-    private const FLAG_RFU2 = 0b00111000;
-    private const FLAG_AT = 0b01000000;
-    private const FLAG_ED = 0b10000000;
-
     public function __construct(string $authData, string $rpIdHash, string $flags, int $signCount, ?AttestedCredentialData $attestedCredentialData, ?AuthenticationExtensionsClientOutputs $extensions)
     {
         $this->rpIdHash = $rpIdHash;
@@ -79,32 +79,32 @@ class AuthenticatorData
 
     public function isUserPresent(): bool
     {
-        return 0 !== (\ord($this->flags) & self::FLAG_UP) ? true : false;
+        return 0 !== (ord($this->flags) & self::FLAG_UP) ? true : false;
     }
 
     public function isUserVerified(): bool
     {
-        return 0 !== (\ord($this->flags) & self::FLAG_UV) ? true : false;
+        return 0 !== (ord($this->flags) & self::FLAG_UV) ? true : false;
     }
 
     public function hasAttestedCredentialData(): bool
     {
-        return 0 !== (\ord($this->flags) & self::FLAG_AT) ? true : false;
+        return 0 !== (ord($this->flags) & self::FLAG_AT) ? true : false;
     }
 
     public function hasExtensions(): bool
     {
-        return 0 !== (\ord($this->flags) & self::FLAG_ED) ? true : false;
+        return 0 !== (ord($this->flags) & self::FLAG_ED) ? true : false;
     }
 
     public function getReservedForFutureUse1(): int
     {
-        return \ord($this->flags) & self::FLAG_RFU1;
+        return ord($this->flags) & self::FLAG_RFU1;
     }
 
     public function getReservedForFutureUse2(): int
     {
-        return \ord($this->flags) & self::FLAG_RFU2;
+        return ord($this->flags) & self::FLAG_RFU2;
     }
 
     public function getSignCount(): int
