@@ -14,53 +14,63 @@ declare(strict_types=1);
 namespace Webauthn;
 
 use Assert\Assertion;
+use function count;
+use function in_array;
 use InvalidArgumentException;
+use RuntimeException;
+use Safe\Exceptions\FilesystemException;
+use function Safe\file_put_contents;
+use function Safe\ksort;
+use function Safe\mkdir;
+use function Safe\rename;
+use function Safe\sprintf;
+use function Safe\tempnam;
+use function Safe\unlink;
 use Symfony\Component\Process\Process;
 
 class CertificateToolbox
 {
     /**
-     * @param array<string> $authenticatorCertificates
-     * @param array<string> $trustedCertificates
+     * @deprecated "This method is deprecated since v3.3 and will be removed en v4.0. Please use Webauthn\CertificateChainChecker\CertificateChainChecker instead"
+     *
+     * @param string[] $authenticatorCertificates
+     * @param string[] $trustedCertificates
      */
     public static function checkChain(array $authenticatorCertificates, array $trustedCertificates = []): void
     {
-        self::checkCertificatesValidity($authenticatorCertificates);
-        self::checkCertificatesValidity($trustedCertificates);
+        if (0 === count($trustedCertificates)) {
+            self::checkCertificatesValidity($authenticatorCertificates, true);
 
-        if (0 === \count($trustedCertificates)) {
             return;
         }
-        $filenames = [];
+        self::checkCertificatesValidity($authenticatorCertificates, false);
 
-        $leafFilename = tempnam(sys_get_temp_dir(), 'webauthn-leaf-');
-        Assertion::string($leafFilename, 'Unable to get a temporary filename');
+        $processArguments = ['-no-CAfile', '-no-CApath'];
 
-        $leafCertificate = array_shift($authenticatorCertificates);
-        $result = file_put_contents($leafFilename, $leafCertificate);
-        Assertion::integer($result, 'Unable to write temporary data');
-        $filenames[] = $leafFilename;
-
-        $processArguments = ['--no-CApath', '--no-CAfile'];
+        $caDirname = self::createTemporaryDirectory();
+        $processArguments[] = '--CApath';
+        $processArguments[] = $caDirname;
 
         foreach ($trustedCertificates as $certificate) {
-            $trustedFilename = tempnam(sys_get_temp_dir(), 'webauthn-trusted-');
-            Assertion::string($trustedFilename, 'Unable to get a temporary filename');
-            $result = file_put_contents($trustedFilename, $certificate, FILE_APPEND);
-            Assertion::integer($result, 'Unable to write temporary data');
-            $result = file_put_contents($trustedFilename, PHP_EOL, FILE_APPEND);
-            Assertion::integer($result, 'Unable to write temporary data');
-            $processArguments[] = '-trusted';
-            $processArguments[] = $trustedFilename;
-            $filenames[] = $trustedFilename;
+            self::prepareCertificate($caDirname, $certificate, 'webauthn-trusted-', '.pem');
         }
+
+        $rehashProcess = new Process(['openssl', 'rehash', $caDirname]);
+        $rehashProcess->run();
+        while ($rehashProcess->isRunning()) {
+            //Just wait
+        }
+        if (!$rehashProcess->isSuccessful()) {
+            throw new InvalidArgumentException('Invalid certificate or certificate chain');
+        }
+
+        $filenames = [];
+        $leafCertificate = array_shift($authenticatorCertificates);
+        $leafFilename = self::prepareCertificate(sys_get_temp_dir(), $leafCertificate, 'webauthn-leaf-', '.pem');
+        $filenames[] = $leafFilename;
+
         foreach ($authenticatorCertificates as $certificate) {
-            $untrustedFilename = tempnam(sys_get_temp_dir(), 'webauthn-untrusted-');
-            Assertion::string($untrustedFilename, 'Unable to get a temporary filename');
-            $result = file_put_contents($untrustedFilename, $certificate, FILE_APPEND);
-            Assertion::integer($result, 'Unable to write temporary data');
-            $result = file_put_contents($untrustedFilename, PHP_EOL, FILE_APPEND);
-            Assertion::integer($result, 'Unable to write temporary data');
+            $untrustedFilename = self::prepareCertificate(sys_get_temp_dir(), $certificate, 'webauthn-untrusted-', '.pem');
             $processArguments[] = '-untrusted';
             $processArguments[] = $untrustedFilename;
             $filenames[] = $untrustedFilename;
@@ -70,45 +80,51 @@ class CertificateToolbox
         array_unshift($processArguments, 'openssl', 'verify');
 
         $process = new Process($processArguments);
-        $process->start();
+        $process->run();
         while ($process->isRunning()) {
-        }
-        foreach ($filenames as $filename) {
-            $result = unlink($filename);
-            Assertion::true($result, 'Unable to delete temporary file');
+            //Just wait
         }
 
+        foreach ($filenames as $filename) {
+            try {
+                unlink($filename);
+            } catch (FilesystemException $e) {
+                continue;
+            }
+        }
+        self::deleteDirectory($caDirname);
+
         if (!$process->isSuccessful()) {
-            throw new InvalidArgumentException('Invalid certificate or certificate chain. Error is: '.$process->getErrorOutput());
+            throw new InvalidArgumentException('Invalid certificate or certificate chain');
         }
     }
 
-    public static function fixPEMStructure(string $certificate): string
+    public static function fixPEMStructure(string $certificate, string $type = 'CERTIFICATE'): string
     {
-        $pemCert = '-----BEGIN CERTIFICATE-----'.PHP_EOL;
+        $pemCert = '-----BEGIN '.$type.'-----'.PHP_EOL;
         $pemCert .= chunk_split($certificate, 64, PHP_EOL);
-        $pemCert .= '-----END CERTIFICATE-----'.PHP_EOL;
+        $pemCert .= '-----END '.$type.'-----'.PHP_EOL;
 
         return $pemCert;
     }
 
-    public static function convertDERToPEM(string $certificate): string
+    public static function convertDERToPEM(string $certificate, string $type = 'CERTIFICATE'): string
     {
         $derCertificate = self::unusedBytesFix($certificate);
 
-        return self::fixPEMStructure(base64_encode($derCertificate));
+        return self::fixPEMStructure(base64_encode($derCertificate), $type);
     }
 
     /**
-     * @param array<string> $certificates
+     * @param string[] $certificates
      *
-     * @return array<string>
+     * @return string[]
      */
-    public static function convertAllDERToPEM(array $certificates): array
+    public static function convertAllDERToPEM(array $certificates, string $type = 'CERTIFICATE'): array
     {
         $certs = [];
         foreach ($certificates as $publicKey) {
-            $certs[] = self::convertDERToPEM($publicKey);
+            $certs[] = self::convertDERToPEM($publicKey, $type);
         }
 
         return $certs;
@@ -117,7 +133,7 @@ class CertificateToolbox
     private static function unusedBytesFix(string $certificate): string
     {
         $certificateHash = hash('sha256', $certificate);
-        if (\in_array($certificateHash, self::getCertificateHashes(), true)) {
+        if (in_array($certificateHash, self::getCertificateHashes(), true)) {
             $certificate[mb_strlen($certificate, '8bit') - 257] = "\0";
         }
 
@@ -125,13 +141,17 @@ class CertificateToolbox
     }
 
     /**
-     * @param array<string> $certificates
+     * @param string[] $certificates
      */
-    private static function checkCertificatesValidity(array $certificates): void
+    private static function checkCertificatesValidity(array $certificates, bool $allowRootCertificate): void
     {
         foreach ($certificates as $certificate) {
             $parsed = openssl_x509_parse($certificate);
             Assertion::isArray($parsed, 'Unable to read the certificate');
+            if (false === $allowRootCertificate) {
+                self::checkRootCertificate($parsed);
+            }
+
             Assertion::keyExists($parsed, 'validTo_time_t', 'The certificate has no validity period');
             Assertion::keyExists($parsed, 'validFrom_time_t', 'The certificate has no validity period');
             Assertion::lessOrEqualThan(time(), $parsed['validTo_time_t'], 'The certificate expired');
@@ -140,7 +160,21 @@ class CertificateToolbox
     }
 
     /**
-     * @return array<string>
+     * @param array<string, mixed> $parsed
+     */
+    private static function checkRootCertificate(array $parsed): void
+    {
+        Assertion::keyExists($parsed, 'subject', 'The certificate has no subject');
+        Assertion::keyExists($parsed, 'issuer', 'The certificate has no issuer');
+        $subject = $parsed['subject'];
+        $issuer = $parsed['issuer'];
+        ksort($subject);
+        ksort($issuer);
+        Assertion::notEq($subject, $issuer, 'Root certificates are not allowed');
+    }
+
+    /**
+     * @return string[]
      */
     private static function getCertificateHashes(): array
     {
@@ -152,5 +186,38 @@ class CertificateToolbox
             '6073c436dcd064a48127ddbf6032ac1a66fd59a0c24434f070d4e564c124c897',
             'ca993121846c464d666096d35f13bf44c1b05af205f9b4a1e00cf6cc10c5e511',
         ];
+    }
+
+    private static function createTemporaryDirectory(): string
+    {
+        $caDir = tempnam(sys_get_temp_dir(), 'webauthn-ca-');
+        if (file_exists($caDir)) {
+            unlink($caDir);
+        }
+        mkdir($caDir);
+        if (!is_dir($caDir)) {
+            throw new RuntimeException(sprintf('Directory "%s" was not created', $caDir));
+        }
+
+        return $caDir;
+    }
+
+    private static function deleteDirectory(string $dirname): void
+    {
+        $rehashProcess = new Process(['rm', '-rf', $dirname]);
+        $rehashProcess->run();
+        while ($rehashProcess->isRunning()) {
+            //Just wait
+        }
+    }
+
+    private static function prepareCertificate(string $folder, string $certificate, string $prefix, string $suffix): string
+    {
+        $untrustedFilename = tempnam($folder, $prefix);
+        rename($untrustedFilename, $untrustedFilename.$suffix);
+        file_put_contents($untrustedFilename.$suffix, $certificate, FILE_APPEND);
+        file_put_contents($untrustedFilename.$suffix, PHP_EOL, FILE_APPEND);
+
+        return $untrustedFilename.$suffix;
     }
 }
