@@ -44,7 +44,7 @@ abstract class PublicKeyCredentialEntity implements JsonSerializable
     }
 
     /**
-     * @return array<string, mixed>
+     * @return mixed[]
      */
     public function jsonSerialize(): array
     {
