@@ -15,6 +15,7 @@ namespace Webauthn;
 
 use Assert\Assertion;
 use JsonSerializable;
+use function Safe\json_decode;
 
 class PublicKeyCredentialParameters implements JsonSerializable
 {
@@ -47,14 +48,13 @@ class PublicKeyCredentialParameters implements JsonSerializable
     public static function createFromString(string $data): self
     {
         $data = json_decode($data, true);
-        Assertion::eq(JSON_ERROR_NONE, json_last_error(), 'Invalid data');
         Assertion::isArray($data, 'Invalid data');
 
         return self::createFromArray($data);
     }
 
     /**
-     * @param array<string, mixed> $json
+     * @param mixed[] $json
      */
     public static function createFromArray(array $json): self
     {
@@ -70,7 +70,7 @@ class PublicKeyCredentialParameters implements JsonSerializable
     }
 
     /**
-     * @return array<string, mixed>
+     * @return mixed[]
      */
     public function jsonSerialize(): array
     {
