@@ -19,6 +19,8 @@ use InvalidArgumentException;
 use JsonSerializable;
 use Ramsey\Uuid\Uuid;
 use Ramsey\Uuid\UuidInterface;
+use function Safe\base64_decode;
+use function Safe\sprintf;
 use Throwable;
 use Webauthn\TrustPath\TrustPath;
 use Webauthn\TrustPath\TrustPathLoader;
@@ -74,7 +76,7 @@ class PublicKeyCredentialSource implements JsonSerializable
     protected $counter;
 
     /**
-     * @param array<string> $transports
+     * @param string[] $transports
      */
     public function __construct(string $publicKeyCredentialId, string $type, array $transports, string $attestationType, TrustPath $trustPath, UuidInterface $aaguid, string $credentialPublicKey, string $userHandle, int $counter)
     {
@@ -161,7 +163,7 @@ class PublicKeyCredentialSource implements JsonSerializable
     }
 
     /**
-     * @param array<string, mixed> $data
+     * @param mixed[] $data
      */
     public static function createFromArray(array $data): self
     {
@@ -175,7 +177,6 @@ class PublicKeyCredentialSource implements JsonSerializable
                 break;
             default: // Kept for compatibility with old format
                 $decoded = base64_decode($data['aaguid'], true);
-                Assertion::string($decoded, 'Invalid AAGUID');
                 $uuid = Uuid::fromBytes($decoded);
         }
 
@@ -197,7 +198,7 @@ class PublicKeyCredentialSource implements JsonSerializable
     }
 
     /**
-     * @return array<string, mixed>
+     * @return mixed[]
      */
     public function jsonSerialize(): array
     {
