@@ -13,8 +13,10 @@ declare(strict_types=1);
 
 namespace Webauthn\TokenBinding;
 
+use function array_key_exists;
 use Assert\Assertion;
 use Base64Url\Base64Url;
+use function Safe\sprintf;
 
 class TokenBinding
 {
@@ -40,7 +42,7 @@ class TokenBinding
     }
 
     /**
-     * @param array<string, mixed> $json
+     * @param mixed[] $json
      */
     public static function createFormArray(array $json): self
     {
@@ -51,7 +53,7 @@ class TokenBinding
             self::getSupportedStatus(),
             sprintf('The member "status" is invalid. Supported values are: %s', implode(', ', self::getSupportedStatus()))
         );
-        $id = \array_key_exists('id', $json) ? Base64Url::decode($json['id']) : null;
+        $id = array_key_exists('id', $json) ? Base64Url::decode($json['id']) : null;
 
         return new self($status, $id);
     }
