@@ -16,7 +16,7 @@ namespace Webauthn\TrustPath;
 final class EmptyTrustPath implements TrustPath
 {
     /**
-     * @return array<string, string>
+     * @return string[]
      */
     public function jsonSerialize(): array
     {
